requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tabulate>=0.9.0
openai>=1.0.0
joblib>=1.3.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tenacity>=8.2.0
python-telegram-bot>=20.0
pulp>=2.7.0
